Fixed getById() problem, added main page, controller->service->dao

+++ src/main/java/com/softserve/team5/dao/implementations/UserDaoImpl.java
@@ -2,52 +2,55 @@ package com.softserve.team5.dao.implementations;
 
 import com.softserve.team5.dao.interfaces.UserDao;
 import com.softserve.team5.entity.User;
-
-
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.time.LocalDate;
 
 @Repository
 @EnableTransactionManagement
 @Transactional
 public class UserDaoImpl implements UserDao {
 
-	private SessionFactory sessionFactory;
+    private final SessionFactory sessionFactory;
 
-	@Override
-	public void create(User entity) {
-		Session session = sessionFactory.getCurrentSession();
-		session.persist(entity);
-	}
+    @Autowired
+    public UserDaoImpl(SessionFactory sessionFactory) {
+        this.sessionFactory = sessionFactory;
+    }
 
-	@Override
-	public void update(User entity) {
-		Session session = sessionFactory.getCurrentSession();
-		session.update(entity);
+    @Override
+    public void create(User entity) {
+        Session session = sessionFactory.getCurrentSession();
+        session.persist(entity);
+    }
 
-	}
+    @Override
+    public void update(User user) {
+        Session session = sessionFactory.getCurrentSession();
+        user.setRegistrationDate(LocalDate.now());
+        session.update(user);
+    }
 
-	@Override
-	public void delete(Long id) {
-		Session session = sessionFactory.getCurrentSession();
-		User user = session.load(User.class, new Long(id));
-		if (user != null) {
-			session.delete(user);
-		}
-	}
+    @Override
+    public void delete(Long id) {
+        Session session = sessionFactory.getCurrentSession();
+        User user = getById(id);
+        if (user != null) session.delete(user);
+    }
 
-	@Override
-	public User getById(Long id) {
-		Session session = sessionFactory.getCurrentSession();
-		User user = session.load(User.class, new Long(id));
-		return user;
-	}
+    @Override
+    public User getById(Long id) {
+        Session session = sessionFactory.getCurrentSession();
+        return session.get(User.class, id);
+    }
 
-	@Override
-	public double getAverengeUsersAge() {
-		return 0;
-	}
+    @Override
+    public double getAverageUsersAge() {
+        return 0;
+    }
 }
